--- src/routes/get/[[parent]]/[what]/load.js
@@ -1,11 +1,0 @@
-/*
-
-terms
-thorpes
-octothorpes
-links
-backlinks
-bookmarks
-all
-
-*/
